requirements.txt
PyOpenGL==3.1.6
pyrr==0.10.3
PyYAML==6.0
numpy==1.24.1
colorama==0.4.6
glfw==2.5.5